requirements.txt
PyYAML
cryptography
